requirements.txt
PyYAML
cryptography
